+++ tests/test_flake8_async.py
@@ -543,6 +543,7 @@ error_codes_ignored_when_checking_transformed_sync_code = {
     "ASYNC127",
     "ASYNC300",
     "ASYNC400",
+    "ASYNC401",
     "ASYNC912",
 }